+++ src/main/java/com/java_school/cinemabot/services/database/DatabaseSessionService.java
@@ -12,3 +12,4 @@ public interface DatabaseSessionService {
     List<Session> getSessionByDate(LocalDate date);
     void saveSessions(List<SessionDTO> sessions);
 }
+
